core/logging: Move `ml::Tag` into its own header and expose a little more API
This lets us drop the `logging_internal.h` header and instead fully drive the logging
configuration from the main `mir::options` system, avoiding a dependency on
boost::program_option in mircore (and, later, avoiding a dependency on libmiral
in mircore)
