+++ src/core/CMakeLists.txt
@@ -13,6 +13,7 @@ add_library(mircore SHARED
     log.cpp
     logging/dumb_console_logger.cpp
     logging/logger.cpp
+    logging/tag.cpp
     report_exception.cpp ${PROJECT_SOURCE_DIR}/include/core/mir/report_exception.h
     ${PROJECT_SOURCE_DIR}/include/core/mir/anonymous_shm_file.h
     ${PROJECT_SOURCE_DIR}/include/core/mir/int_wrapper.h
@@ -29,6 +30,7 @@ add_library(mircore SHARED
     ${PROJECT_SOURCE_DIR}/include/core/mir/log.h
     ${PROJECT_SOURCE_DIR}/include/core/mir/logging/dumb_console_logger.h
     ${PROJECT_SOURCE_DIR}/include/core/mir/logging/logger.h
+    ${PROJECT_SOURCE_DIR}/include/core/mir/logging/tag.h
     ${PROJECT_SOURCE_DIR}/include/core/mir/shm_file.h
     ${PROJECT_SOURCE_DIR}/include/core/mir/depth_layer.h
     ${PROJECT_SOURCE_DIR}/include/core/mir_toolkit/common.h
@@ -37,18 +39,10 @@ add_library(mircore SHARED
     ${PROJECT_SOURCE_DIR}/include/core/mir/input/mousekeys_keymap.h
 )
 
-target_link_libraries(mircore
-  PRIVATE
-    Boost::program_options
-)
-
 target_include_directories(mircore
   PUBLIC
     ${PROJECT_SOURCE_DIR}/include/core
-  PRIVATE
-    ${PROJECT_SOURCE_DIR}/src/include
 )
-
 
 set(symbol_map ${CMAKE_CURRENT_SOURCE_DIR}/symbols.map)
